+++ app/src/main/java/com/shuyun/qapp/bean/BannerBean.java
@@ -12,7 +12,7 @@ public class BannerBean {
     private String remark;//描述
     private String picture;//主图
     private String url;//跳转的地址
-    private int type;//类型 1、外部链接;2、内部链接;3、内部功能
+    private Long type;//类型 1、外部链接;2、内部链接;3、内部功能
     private String h5Url; //答题url
     private String action;
     private String content;
@@ -53,7 +53,7 @@ public class BannerBean {
         return url;
     }
 
-    public int getType() {
+    public Long getType() {
         return type;
     }
